refactor(member): reindent codebase to 2 spaces
