+++ src/main/java/com/kindtalk/server/member/domain/Member.java
@@ -11,28 +11,28 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Member {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    private String email;
-    private String password;
-    private String userName;
-    private String nickName;
+  private String email;
+  private String password;
+  private String userName;
+  private String nickName;
+  
+  @Enumerated(EnumType.STRING)
+  private Role role;
 
-    @Enumerated(EnumType.STRING)
-    private Role role;
+  public Member(String email, String password, String userName, String nickName, Role role) {
+    this.email = email;
+    this.password = password;
+    this.userName = userName;
+    this.nickName = nickName;
+    this.role = role;
+  }
 
-    public Member(String email, String password, String userName, String nickName, Role role) {
-        this.email = email;
-        this.password = password;
-        this.userName = userName;
-        this.nickName = nickName;
-        this.role = role;
-    }
-
-    public void updateInfo(String userName, String nickName) {
-        this.userName = userName;
-        this.nickName = nickName;
-    }
+  public void updateInfo(String userName, String nickName) {
+    this.userName = userName;
+    this.nickName = nickName;
+  }
 }
