+++ src/main/java/com/kindtalk/server/member/dto/MemberResponse.java
@@ -4,17 +4,17 @@ import com.kindtalk.server.member.domain.Member;
 import com.kindtalk.server.member.role.Role;
 
 public record MemberResponse(
-    String email,
-    String userName,
-    String nickName,
-    Role role) {
+  String email,
+  String userName,
+  String nickName,
+  Role role) {
 
-    public static MemberResponse of(Member member) {
-        return new MemberResponse(
-            member.getEmail(),
-            member.getUserName(),
-            member.getNickName(),
-            member.getRole()
-        );
-    }
+  public static MemberResponse of(Member member) {
+    return new MemberResponse(
+      member.getEmail(),
+      member.getUserName(),
+      member.getNickName(),
+      member.getRole()
+    );
+  }
 }
