+++ src/main/java/com/kindtalk/server/member/dto/MemberUpdateRequest.java
@@ -3,9 +3,9 @@ package com.kindtalk.server.member.dto;
 import jakarta.validation.constraints.NotBlank;
 
 public record MemberUpdateRequest(
-    @NotBlank
-    String userName,
+  @NotBlank
+  String userName,
 
-    @NotBlank
-    String nickName) {
+  @NotBlank
+  String nickName) {
 }
