+++ src/main/java/com/kindtalk/server/member/repository/MemberRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
-    Optional<Member> findByEmail(String email);
+  Optional<Member> findByEmail(String email);
 
-    Boolean existsByEmail(String email);
+  Boolean existsByEmail(String email);
 }
