+++ src/main/java/com/kindtalk/server/member/role/Role.java
@@ -1,5 +1,5 @@
 package com.kindtalk.server.member.role;
 
 public enum Role {
-    TEACHER, MEMBER
+  TEACHER, MEMBER
 }
